requirements.txt
mcp>=1.0.0
kroger-api
python-dotenv
